telegram2fido/requirements.txt
python-telegram-bot>=21.0
telethon>=1.34
